The name of the store is now stored in the store
Using a hashmap made the ordering of the stores inconsistent,
and we need to keep order for preferences.

So now its a vector, which to search for a given store should
not be slower, as there should not be many stores (10 max?).

One think I dont like is that the name before was external to
the store, and now is internal, and somehow I dont see it
belongs there.

We could extract it, creating a vec of pairs(str, Store).. But as I
did have problems before with accesing the name of the store,
lets try to have it internally.